Add files via upload
In this project, you will act as a data visualization developer at Yahoo Finance! You will be helping the "Netflix Stock Profile" team visualize the Netflix stock data. In finance, a stock profile is a series of studies, visualizations, and analyses that dive into different aspects a publicly traded company's data.

For the purposes of the project, you will only visualize data for the year of 2017. Specifically, you will be in charge of creating the following visualizations:

The distribution of the stock prices for the past year
Netflix's earnings and revenue in the last four quarters
The actual vs. estimated earnings per share for the four quarters in 2017
A comparison of the Netflix Stock price vs the Dow Jones Industrial Average price in 2017
Note: We are using the Dow Jones Industrial Average to compare the Netflix stock to the larter stock market. Learn more about why the Dow Jones Industrial Average is a general reflection of the larger stock market here.